% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_calculation.R
\name{Bf_calculation}
\alias{Bf_calculation}
\title{BF function based on Dienes & McLatchie, 2018}
\usage{
Bf_calculation(sd, obtained, dfdata, meanoftheory, sdtheory, tail = 2)
}
\arguments{
\item{sd}{SE of data}
}
\description{
Modified so that H1 is represented by normal distribution (rather than t), hence there is no 'dftheory' argument
}
\section{notes}{
 this function is based on the Dienes & Mclatchie (2018) calculator
}

