% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_calculation.R
\name{Bf_normal}
\alias{Bf_normal}
\title{Function to calculate Bayes factor with normal prior based on Dienes & McLatchie, 2018}
\usage{
Bf_normal(sd, obtained, dfdata, mean_of_theory, sd_of_theory, tail = c(1, 2))
}
\arguments{
\item{sd}{numeric. SE of data}

\item{obtained}{numeric.}

\item{dfdata}{integer.}

\item{mean_of_theory}{numeric.}

\item{sd_of_theory}{numeric.}

\item{tail}{integer.}
}
\value{
The function returns...
}
\description{
Modified so that H1 is represented by normal distribution (rather than t), hence there is no 'dftheory' argument
}
\section{notes}{
 this function is based on the Dienes & Mclatchie (2018) calculator
}

