% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_samplesize.R
\name{Bf_samplesize}
\alias{Bf_samplesize}
\title{title}
\usage{
Bf_samplesize(
  Bf_calculation,
  n_start,
  n_end,
  step,
  sd1,
  sd2 = NULL,
  threshold = c(3, 6, 10),
  tpr = c(0.5, 0.8, 0.9, 0.95),
  ...
)
}
\description{
The function assumes equal group sizes for between-subject designs.
}
