% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_samplesize.R
\name{Bf_samplesize_uniform}
\alias{Bf_samplesize_uniform}
\title{Function to calculate the samplesize with uniform prior
 
 This function ...
 
 @param upper numeric.
 @param sd1 numeric.
 @param sd2 numeric.
 @param tail integer.
 @param threshold integer.
 @param tpr numeric.
 @param n_start integer.
 @param n_end integer.
 @param step integer.
 
 @return The function returns ...}
\usage{
Bf_samplesize_uniform(
  upper,
  sd1,
  sd2 = NULL,
  tail = c(1, 2),
  threshold = c(3, 6, 10),
  tpr = c(0.5, 0.8, 0.9, 0.95),
  n_start = 5,
  n_end = 100,
  step = 1
)
}
\description{
Function to calculate the samplesize with uniform prior
 
 This function ...
 
 @param upper numeric.
 @param sd1 numeric.
 @param sd2 numeric.
 @param tail integer.
 @param threshold integer.
 @param tpr numeric.
 @param n_start integer.
 @param n_end integer.
 @param step integer.
 
 @return The function returns ...
}
