% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_calculation.R
\name{Bf_uniform}
\alias{Bf_uniform}
\title{Function to calculate Bayes factor with uniform prior based on Dienes & McLatchie, 2018}
\usage{
Bf_uniform(sd, obtained, dfdata, lower, upper)
}
\arguments{
\item{sd}{numeric. SE of data}

\item{obtained}{numeric.}

\item{dfdata}{integer.}

\item{lower}{numeric.}

\item{upper}{numeric.}
}
\value{
The function returns...
}
\description{
description of the function
}
\section{notes}{
 this function is based on the Dienes & Mclatchie (2018) calculator
}

