% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_se.R
\name{calculate_se}
\alias{calculate_se}
\title{Function to calculate SE for between and within designs}
\usage{
calculate_se(sd1, sd2 = NULL, n)
}
\arguments{
\item{sd1}{numeric.}

\item{sd2}{numeric.}

\item{n}{integer.}
}
\value{
The function returns...
}
\description{
description
}
