% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_simulation.R
\name{fixed_BF}
\alias{fixed_BF}
\title{Function to calculate Bayes factor with fixed stopping}
\usage{
fixed_BF(sample, sd_of_theory, tail = c(1, 2), Bf_calculation)
}
\arguments{
\item{sample}{tibble.}

\item{sd_of_theory}{numeric.}

\item{tail}{integer.}

\item{Bf_calculation}{function object.}
}
\value{
The function returns
}
\description{
This function ...
}
