% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_simulation.R
\name{optional_BF}
\alias{optional_BF}
\title{Function to calculate Bayes factor with optional stopping}
\usage{
optional_BF(sample, threshold, sd_of_theory, tail = c(1, 2), Bf_calculation)
}
\arguments{
\item{sample}{tibble.}

\item{threshold}{integer.}

\item{sd_of_theory}{numeric.}

\item{tail}{integer.}

\item{Bf_calculation}{function object.}
}
\value{
The function returns
}
\description{
The function ...
}
