% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_simulation.R
\name{simulate_Bf}
\alias{simulate_Bf}
\title{Function simulating studies and Bayesian sequential analyses}
\usage{
simulate_Bf(
  sd_of_theory,
  sd1,
  sd2 = NULL,
  correlation = 0,
  threshold = c(3, 6, 10),
  stopping_rule = c("optional", "fixed"),
  n = 100,
  iterations = 5,
  Bf_calculation = c("Bf_normal", "Bf_cauchy"),
  tail = c(1, 2),
  true_effect = TRUE
)
}
\arguments{
\item{sd_of_theory}{numeric.}

\item{sd1}{numeric.}

\item{sd2}{numeric.}

\item{correlation}{numeric.}

\item{threshold}{integer.}

\item{stopping_rule}{character.}

\item{n}{integer.}

\item{iterations}{integer.}

\item{Bf_calculation}{function object.}

\item{tail}{integer.}

\item{true_effect}{logical.}
}
\value{
The function returns a tibble with two columns: `id` that contains
  the unique identifier of the given iteration, and `hit` which contains whether
  the resulting Bayes factor was larger than the specified threshold (`2`),
  inconclusive (`0`), or smaller than 1 / threshold (`1`).
}
\description{
This function...
}
