% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_calculation.R
\name{Bf_cauchy}
\alias{Bf_cauchy}
\title{Function to calculate Bayes factor with Cauchy prior based on Dienes & McLatchie, 2018}
\usage{
Bf_cauchy(sd, obtained, dfdata, mean_of_theory, sd_of_theory, tail = c(1, 2))
}
\arguments{
\item{sd}{numeric. SE of data}

\item{obtained}{numeric.}

\item{dfdata}{integer.}

\item{mean_of_theory}{numeric.}

\item{sd_of_theory}{numeric.}

\item{tail}{integer.}
}
\value{
The function returns...
}
\description{
description of the function
}
\section{notes}{
 this function is based on the Dienes & Mclatchie (2018) calculator
}

