% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_samplesize.R
\name{Bf_samplesize}
\alias{Bf_samplesize}
\title{title}
\usage{
Bf_samplesize(
  mean_of_theory,
  sd_of_theory,
  sd1,
  sd2 = NULL,
  tail = c(1, 2),
  threshold = c(3, 6, 10),
  tpr = c(0.5, 0.8, 0.9, 0.95),
  design = c("between", "within"),
  n_start = 5,
  n_end = 100,
  step = 1
)
}
\description{
The function assumes equal group sizes for between-subject designs.
}
