% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_calculation.R
\name{Bf_uniform}
\alias{Bf_uniform}
\title{Calculate Bayes factor with uniform prior}
\usage{
Bf_uniform(sd, obtained, dfdata, lower, upper)
}
\arguments{
\item{sd}{numeric. Calculated SE of the observed effect. See \code{\link{calculate_se}}.}

\item{obtained}{numeric. The observed effect in raw units.}

\item{dfdata}{integer. The degrees of freedom of the sample. See \code{\link{calculate_df}}.}

\item{lower}{numeric. The smallest predicted effect size by H1.}

\item{upper}{numeric. The largest predicted effect size by H1.}
}
\value{
The function returns the Bayes factor (evidence for H1 over H0) as a single
numeric value.
}
\description{
This function calculates the Bayes factor, i.e. the evidence for H1 over H0.
The function models the predictions of H1 with a uniform distribution, and
assumes that the data generation process (i.e. the likelihood function) follows
a t-distribution. The function is based on the Dienes (2008) calculator.
}
\section{Note}{
 Please set either `lower` or `upper` equal to 0 in order to
  model a directional H1.
}

\examples{
# One-tailed test with the observed effect in the same direction as H1 predicted
SampleSizeBf::Bf_uniform(sd = 0.4, obtained = 1, dfdata = 98, lower = 0,
                        upper = 1.5, tail = 1)
# One-tailed test with the observed effect in the opposite direction as H1 predicted
SampleSizeBf::Bf_uniform(sd = 0.4, obtained = -0.3, dfdata = 98, lower = 0,
                        upper = 1.5, tail = 1)

}
