% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_simulation.R
\name{generate_sample}
\alias{generate_sample}
\title{Function for generating samples}
\usage{
generate_sample(n, sd_of_theory, sd1, sd2 = NULL, correlation = correlation)
}
\arguments{
\item{n}{integer.}

\item{sd_of_theory}{numeric.}

\item{sd1}{numeric.}

\item{sd2}{numeric.}

\item{correlation}{numeric.}
}
\value{
The function returns a tibble with ...
}
\description{
This function
}
