% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_simulation.R
\name{simulate_BF}
\alias{simulate_BF}
\title{Function simulating studies and Bayesian sequential analyses}
\usage{
simulate_BF(
  sd_of_theory,
  sd1,
  sd2 = NULL,
  correlation = 0,
  threshold = 3,
  stopping_rule = c("optional", "fixed"),
  n_start = 5,
  n_end = 100,
  step = 10,
  iterations = 5
)
}
\arguments{
\item{mean_exp}{Mean in the experimental condition}
}
\value{
The function returns four values
}
\description{
This function...
}
