% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bf_simulation.R
\name{simulate_BF}
\alias{simulate_BF}
\title{Function simulating studies and Bayesian sequential analyses}
\usage{
simulate_BF(
  N,
  mean_exp,
  SD_exp,
  mean_con,
  SD_con,
  correlation = 0,
  sdtheory,
  cutoff = 3,
  stopping_rule = c("optional", "fixed"),
  steps = 10,
  iterations = 1000
)
}
\arguments{
\item{N}{sample size}

\item{mean_exp}{Mean in the experimental condition}
}
\value{
The function returns four values
}
\description{
This function...
}
